Changed things in both services, dbcontext, and migrations configuration.
