--- ShoppingList/ShoppingList/Models/ShoppingListModels/ShoppingListViewModel.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Web;
-
-namespace ShoppingList.Models.ShoppingListModels.ViewModels
-{
-    public class ShoppingListViewModel
-    {
-        public string ListName { get; set; }
-    }
-}
