+++ ShoppingList/ShoppingList/Models/ShoppingListModels/ViewModels/ListOfListsViewModel.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Web;
+
+namespace ShoppingList.Models.ShoppingListModels
+{
+    public class ListOfListsViewModel
+    {
+        [Key]
+        [Required]
+        public int ListId { get; set; }
+
+        [Required]
+        [DisplayName("List Name")]
+        public string ListName { get; set; }
+
+        [Required]
+        public string Color { get; set; }
+
+    }
+}
