+++ ShoppingList/ShoppingList/ShoppingList.csproj
@@ -184,15 +184,15 @@
     <Compile Include="Models\AccountViewModels.cs" />
     <Compile Include="Models\IdentityModels.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
-    <Compile Include="Models\ShoppingListModels\IdentityModel.cs" />
-    <Compile Include="Models\ShoppingListModels\ListOfListsViewModel.cs" />
-    <Compile Include="Models\ShoppingListModels\ShoppingListDetailViewModel.cs" />
-    <Compile Include="Models\ShoppingListModels\ShoppingListEntity.cs" />
-    <Compile Include="Models\ShoppingListModels\ShoppingListItem.cs" />
-    <Compile Include="Models\ShoppingListModels\ShoppingListItemCreateViewModel.cs" />
-    <Compile Include="Models\ShoppingListModels\ShoppingListNote.cs" />
-    <Compile Include="Models\ShoppingListModels\ShoppingListCreateViewModel.cs" />
-    <Compile Include="Models\ShoppingListModels\ShoppingListViewModel.cs" />
+    <Compile Include="Models\ShoppingListModels\DataModels\IdentityModel.cs" />
+    <Compile Include="Models\ShoppingListModels\ViewModels\ListOfListsViewModel.cs" />
+    <Compile Include="Models\ShoppingListModels\ViewModels\ShoppingListItemsViewModel.cs" />
+    <Compile Include="Models\ShoppingListModels\DataModels\ShoppingListEntity.cs" />
+    <Compile Include="Models\ShoppingListModels\DataModels\ShoppingListItemEntity.cs" />
+    <Compile Include="Models\ShoppingListModels\ViewModels\ShoppingListItemCreateViewModel.cs" />
+    <Compile Include="Models\ShoppingListModels\DataModels\ShoppingListNoteEntity.cs" />
+    <Compile Include="Models\ShoppingListModels\ViewModels\ShoppingListCreateViewModel.cs" />
+    <Compile Include="Models\ShoppingListModels\ViewModels\ShoppingListViewModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ShoppingListItemService.cs" />
     <Compile Include="ShoppingListService.cs" />
